refactor(core): extract cockpit workflow owner
