+++ crates/tokmd-core/src/workflows/mod.rs
@@ -2,6 +2,8 @@
 
 #[cfg(feature = "analysis")]
 mod analyze;
+#[cfg(feature = "cockpit")]
+mod cockpit;
 mod diff;
 mod export;
 mod lang;
@@ -13,6 +15,10 @@ pub use analyze::{
 };
 #[cfg(all(test, feature = "analysis"))]
 pub(crate) use analyze::{parse_analysis_preset, parse_effort_request};
+#[cfg(feature = "cockpit")]
+pub use cockpit::cockpit_workflow;
+#[cfg(all(test, feature = "cockpit"))]
+pub(crate) use cockpit::parse_cockpit_range_mode;
 pub use diff::diff_workflow;
 pub use export::{export_workflow, export_workflow_from_inputs};
 pub use lang::{lang_workflow, lang_workflow_from_inputs};
